backend/requirements.txt
PyMuPDF>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
boto3>=1.34.0
spacy>=3.7.0
nltk>=3.8.0
# spaCy model (install separately):
#   python -m spacy download en_core_web_sm
